% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_spiralbar.R
\docType{data}
\name{geom_cascade_spiral}
\alias{geom_cascade_spiral}
\alias{GeomCascadeSpiral}
\title{Cascade Spiral and Bar Geom}
\usage{
geom_cascade_spiral(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by `aes()` or `aes_()`. If specified and `inherit.aes = TRUE` (the default), it is combined with the default mapping at the top level of the plot. You must supply `mapping` if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options: If `NULL`, the default, the data is inherited from the plot data as specified in the call to `ggplot`.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{...}{Other arguments passed on to `layer()`. These are often aesthetics, used to set an aesthetic to a fixed value, like `color = 'red'` or `size = 3`. They may also be parameters to the paired stat.}

\item{na.rm}{If `FALSE` (the default), removes missing values with a warning. If `TRUE` silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends? `NA`, the default, includes if any aesthetics are mapped. `FALSE` never includes, and `TRUE` always includes.}

\item{inherit.aes}{If `FALSE`, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification.}
}
\value{
A ggplot2 layer that can be added to a plot.
}
\description{
A custom ggplot2 geom for drawing a combination of a spiral (polyline) and a bar (segment), useful for cascade or spiral visualizations.
}
\examples{
# Example usage (requires appropriate data with 'part' column)
# ggplot(data, aes(x, y, color = color, size = size, part = part)) +
#   geom_cascade_spiral()

}
\seealso{
[ggplot2::layer], [ggplot2::Geom]
}
\keyword{datasets}
